@@ -1,6 +1,8 @@
-require_relative 'boot'
+# frozen_string_literal: true
 
-require 'rails/all'
+require_relative "boot"
+
+require "rails/all"
 
 # TODO : add missing dep to decidim-initiatives/lib/decidim/initiatives/engine.rb
 # require "wicked_pdf"
@@ -14,12 +16,9 @@ module DevelopmentApp
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.2
     config.time_zone = "Europe/Paris"
-    config.i18n.load_path += Dir[Rails.root.join('config/locales/**/*.yml').to_s]
-
+    config.i18n.load_path += Dir[Rails.root.join("config/locales/**/*.yml").to_s]
     # This needs to be set for correct images URLs in emails
-    if Rails.application.secrets.dig(:asset_host).present?
-      config.action_mailer.asset_host = "https://#{Rails.application.secrets.dig(:asset_host)}/"
-    end
+    config.action_mailer.asset_host = "https://#{Rails.application.secrets[:asset_host]}/" if Rails.application.secrets[:asset_host].present?
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
