@@ -2,15 +2,15 @@
 
 if Rails.application.secrets.dig(:scaleway, :id).present? && Rails.env.production?
   CarrierWave.configure do |config|
-    config.fog_provider = 'fog/aws'
+    config.fog_provider = "fog/aws"
     config.fog_credentials = {
-      provider: 'AWS',
+      provider: "AWS",
       aws_access_key_id: Rails.application.secrets.dig(:scaleway, :id),
       aws_secret_access_key: Rails.application.secrets.dig(:scaleway, :token),
       aws_signature_version: 4,
-      region: 'fr-par',
-      host: 's3.fr-par.scw.cloud',
-      endpoint: 'https://s3.fr-par.scw.cloud',
+      region: "fr-par",
+      host: "s3.fr-par.scw.cloud",
+      endpoint: "https://s3.fr-par.scw.cloud",
       enable_signature_v4_streaming: false
     }
     config.storage = :fog
@@ -19,12 +19,10 @@ if Rails.application.secrets.dig(:scaleway, :id).present? && Rails.env.productio
     # config.fog_public = false # optional, defaults to true
     config.fog_directory = Rails.application.secrets.dig(:scaleway, :bucket_name)
     config.fog_attributes = {
-      'Cache-Control' => "max-age=#{365.day.to_i}",
-      'X-Content-Type-Options' => 'nosniff'
+      "Cache-Control" => "max-age=#{365.days.to_i}",
+      "X-Content-Type-Options" => "nosniff"
     }
-    if Rails.application.secrets.dig(:asset_host).present?
-      config.asset_host = "https://#{Rails.application.secrets.dig(:asset_host)}"
-    end
+    config.asset_host = "https://#{Rails.application.secrets[:asset_host]}/" if Rails.application.secrets[:asset_host].present?
   end
 else
   CarrierWave.configure do |config|
