@@ -1,6 +1,8 @@
-%w[
+# frozen_string_literal: true
+
+%w(
   .ruby-version
   .rbenv-vars
   tmp/restart.txt
   tmp/caching-dev.txt
-].each { |path| Spring.watch(path) }
+).each { |path| Spring.watch(path) }
