@@ -1,11 +1,12 @@
 # frozen_string_literal: true
+
 # This migration comes from decidim (originally 20170713131206)
 
 class AddAdminToUsers < ActiveRecord::Migration[5.1]
   def up
     add_column :decidim_users, :admin, :boolean, null: false, default: false
 
-    execute <<~SQL
+    execute <<~SQL.squish
       UPDATE decidim_users
       SET admin = true
       WHERE roles @> '{admin}'
